requirements.txt
django==5.2.3
google-cloud-texttospeech
gunicorn
pillow
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2